Move resolver::Resolve into a new resolver::resolve module
